+++ CHANGELOG.md
@@ -81,6 +81,7 @@ The document follows the conventions described in [“Keep a CHANGELOG”](http:
 - `SVG.Text.textPath()` returns only the first textpath child
 - `attr()` excepts array now to get multiple values at once
 - `SVG.Text.rebuild()` now takes every font-size into account (#512)
+- `fill()` and `stroke()` return the fill and stroke attribute when called as getter (#789)
 
 ### Fixed
 - fixed a bug in clipping and masking where empty nodes persists after removal -> __TODO!__